revert back to the original setTag type
git-svn-id: svn://orca.physics.unc.edu/Orca@11871 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/MJDTestCryoVac/ORMJDTestCryostats/ORMJDTestCryoView.h
@@ -18,9 +18,9 @@
 //-------------------------------------------------------------
 #import "ORMJDPumpCartView.h"
 @interface ORMJDTestCryoView : ORMJDPumpCartView {
-	NSInteger tag;
+	int tag;
 }
-- (NSInteger) tag;
-- (void) setTag:(NSInteger)aValue;
+- (int) tag;
+- (void) setTag:(int)aValue;
 - (void) drawString:(NSString*)aString atPoint:(NSPoint)aPoint;
 @end
